% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor.R
\name{circamean}
\alias{circamean}
\title{Mean of circular quantities}
\usage{
circamean(x, ...)
}
\arguments{
\item{x}{numeric vector representing time-of-day in hours}
}
\value{
a numeric of time-of-day in hours (0-24)
}
\description{
Calculates the mean of
circular quantities for 24hr period phases.
}
\author{
Matthew Carlucci
}
