% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor.R
\name{circdist}
\alias{circdist}
\title{Acrophase distance calculation}
\usage{
circdist(acro1, acro2, per = 24)
}
\arguments{
\item{acro1}{vector of acrophases with period=per}

\item{acro2}{vector of acrophases with period=per}

\item{per}{numeric length 1. Period the acrophase corresponds to.}
}
\value{
shortest distance between pairs of acro1 and acro2
}
\description{
Determine the shortest distance along the unit circle between the two
acrophases (minimum arc-length).
}
\author{
Matthew Carlucci
}
