% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor.R
\name{cosinor-utils}
\alias{cosinor-utils}
\alias{sincos2amp}
\alias{sincos2acro}
\alias{ampacro2cos}
\alias{ampacro2sin}
\alias{acro2complex}
\title{Cosinor Utility Functions}
\usage{
sincos2amp(sin, cos)

sincos2acro(sin, cos, per = 24)

ampacro2cos(amp, acro, per = 24)

ampacro2sin(amp, acro, per = 24)

acro2complex(acro, per = 24)
}
\arguments{
\item{sin}{sin beta coefficient from a cosinor model}

\item{cos}{cos beta coefficient from a cosinor model}

\item{per}{period of oscillation that was fitted (default is 24)}

\item{amp}{numeric, amplitude of a cosinor model}

\item{acro}{numeric, acrophase of a cosinor model}
}
\value{
The amplitude/acrophase of the oscillation or their corresponding linear form coefficients
}
\description{
Applying some useful simple transformations for working with
linear and nonlinear form cosinor parameters.
}
\section{Functions}{
\itemize{
\item \code{sincos2amp()}: Get amplitude from betas

\item \code{sincos2acro()}: Get acrophase from betas

\item \code{ampacro2cos()}: Get cosinor term from acrophase/amplitude

\item \code{ampacro2sin()}: Get sin term from acrophase/amplitude

\item \code{acro2complex()}: Set acrophase as complex for circular mean and variance calculations

}}
\note{
The sin/cos in this context refer to
coefficients (often denoted as beta/gamma) in literature (e.g. Cornelissen 2014).
}
\references{
Cornélissen, G. (2014). Cosinor-Based Rhythmometry.
\emph{Theoretical Biology and Medical Modeling}, \emph{11}, Article 16.
}
\author{
Matthew Carlucci
}
