% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor.R
\name{cosinor_pred}
\alias{cosinor_pred}
\title{Predicted values given the parameters of any cosinor model}
\usage{
cosinor_pred(time, A, acro, mesor = 0, per = 24)
}
\arguments{
\item{time}{numeric, time values to predict}

\item{A}{numeric, cosinor amplitude}

\item{acro}{numeric, cosinor acrophase}

\item{mesor}{numeric, cosinor mesor}

\item{per}{numeric, cosinor period}
}
\description{
Predicted values given the parameters of any cosinor model
}
