% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor.R
\name{discoPMcos}
\alias{discoPMcos}
\title{Fast Multi-Feature Population-Mean Cosinor}
\usage{
discoPMcos(x, period = 24, idname = NULL, alpha = 0.05)
}
\arguments{
\item{x}{list, cosinor data.frames from discoODAs(method="CS")$CS each with
identical number of rows and rows matched by feature. If x is a data.frame it is instead assumed
to be a single feature where rows represent subjects.}
}
\value{
data.frame containing the following resulting columns of statistics with one row per feature:
\item{\code{F}}{F statistic of the zero amplitude test}
\item{\code{df1}}{Model degrees of freedom}
\item{\code{df2}}{Residual degrees of freedom}
\item{\code{pvalue}}{P value of the F-test}
\item{\code{qvalue}}{FDR q-value across all population mean tests (i.e. \code{p.adjust(pvalue,method = "fdr")}}
\item{\code{MESOR}}{Midline Estimating Statistic Of Rhythm}
\item{\code{amplitude}}{Population-mean amplitude estimate}
\item{\code{acrophase}}{Population-mean acrophase estimate}
\item{\code{MESOR_l}}{MESOR lower CI}
\item{\code{MESOR_u}}{MESOR upper CI}
\item{\code{amp_l}}{Amplitude lower CI}
\item{\code{amp_u}}{Amplitude upper CI}
\item{\code{acro_l}}{Acrophase lower CI}
\item{\code{acro_u}}{Acrophase upper CI}
}
\description{
Applies the population-mean cosinor to a set of cosinor results across many features
and many individuals to obtain the related statistics for each feature. Vectorized
for performance benefits.
Expects the output of discoODAs(se,method="CS")$CS cosinor implementation as input,
however, any cosinor implementation can be used which provides sincoef/coscoef/mesor columns.
}
\references{
Cornélissen, G. (2014). Cosinor-Based Rhythmometry.
\emph{Theoretical Biology and Medical Modeling}, \emph{11}, Article 16.

Bingham, C., Arbogast, B., Guillaume Cornélissen, G., Lee, J.K. &
Halberg, F. (1982). Inferential Statistical Methods for Estimating and Comparing
Cosinor Parameters. \emph{Chronobiologia}, \emph{9(4)}, 397-439.
}
