% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor.R
\name{rsqGcos}
\alias{rsqGcos}
\title{Get R squared for Cosinor fits}
\usage{
rsqGcos(pmr, dat)
}
\arguments{
\item{pmr}{data.frame, population-mean results from discoPMcos, with roi and measure columns used for joining}

\item{dat}{data.frame original data used for S-cosinor, with roi and measure columns used for joining}
}
\value{
the original pmr data.frame with two additional columns
\item{\code{rsq}}{Unnormalized R2 - Referred to as R2un throughout the study}
\item{\code{rsqdm}}{Demeaned R2 - Referred to as R2dm throughout the study}
}
\description{
Joins together cosinor results (e.g. discoPMcos) and the original input data
to calculate the two R squared statistics.
}
